+++ lib/kb_ObjectUtilities/kb_ObjectUtilitiesClient.pm
@@ -324,9 +324,9 @@ data_obj_ref is a string
  
 
 
-=head2 KButil_delete_ws_objects
+=head2 KButil_hide_ws_objects
 
-  $return = $obj->KButil_delete_ws_objects($params)
+  $return = $obj->KButil_hide_ws_objects($params)
 
 =over 4
 
@@ -335,16 +335,16 @@ data_obj_ref is a string
 =begin html
 
 <pre>
-$params is a kb_ObjectUtilities.KButil_delete_ws_objects_Params
-$return is a kb_ObjectUtilities.KButil_delete_ws_objects_Output
-KButil_delete_ws_objects_Params is a reference to a hash where the following keys are defined:
+$params is a kb_ObjectUtilities.KButil_hide_ws_objects_Params
+$return is a kb_ObjectUtilities.KButil_hide_ws_objects_Output
+KButil_hide_ws_objects_Params is a reference to a hash where the following keys are defined:
 	workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 	object_types has a value which is a reference to a list where each element is a string
 	verbose has a value which is a kb_ObjectUtilities.bool
-	delete_all has a value which is a kb_ObjectUtilities.bool
+	hide_all has a value which is a kb_ObjectUtilities.bool
 workspace_name is a string
 bool is an int
-KButil_delete_ws_objects_Output is a reference to a hash where the following keys are defined:
+KButil_hide_ws_objects_Output is a reference to a hash where the following keys are defined:
 	report_name has a value which is a kb_ObjectUtilities.data_obj_name
 	report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 data_obj_name is a string
@@ -356,16 +356,16 @@ data_obj_ref is a string
 
 =begin text
 
-$params is a kb_ObjectUtilities.KButil_delete_ws_objects_Params
-$return is a kb_ObjectUtilities.KButil_delete_ws_objects_Output
-KButil_delete_ws_objects_Params is a reference to a hash where the following keys are defined:
+$params is a kb_ObjectUtilities.KButil_hide_ws_objects_Params
+$return is a kb_ObjectUtilities.KButil_hide_ws_objects_Output
+KButil_hide_ws_objects_Params is a reference to a hash where the following keys are defined:
 	workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 	object_types has a value which is a reference to a list where each element is a string
 	verbose has a value which is a kb_ObjectUtilities.bool
-	delete_all has a value which is a kb_ObjectUtilities.bool
+	hide_all has a value which is a kb_ObjectUtilities.bool
 workspace_name is a string
 bool is an int
-KButil_delete_ws_objects_Output is a reference to a hash where the following keys are defined:
+KButil_hide_ws_objects_Output is a reference to a hash where the following keys are defined:
 	report_name has a value which is a kb_ObjectUtilities.data_obj_name
 	report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 data_obj_name is a string
@@ -382,7 +382,7 @@ data_obj_ref is a string
 
 =cut
 
- sub KButil_delete_ws_objects
+ sub KButil_hide_ws_objects
 {
     my($self, @args) = @_;
 
@@ -391,7 +391,7 @@ data_obj_ref is a string
     if ((my $n = @args) != 1)
     {
 	Bio::KBase::Exceptions::ArgumentValidationError->throw(error =>
-							       "Invalid argument count for function KButil_delete_ws_objects (received $n, expecting 1)");
+							       "Invalid argument count for function KButil_hide_ws_objects (received $n, expecting 1)");
     }
     {
 	my($params) = @args;
@@ -399,40 +399,40 @@ data_obj_ref is a string
 	my @_bad_arguments;
         (ref($params) eq 'HASH') or push(@_bad_arguments, "Invalid type for argument 1 \"params\" (value was \"$params\")");
         if (@_bad_arguments) {
-	    my $msg = "Invalid arguments passed to KButil_delete_ws_objects:\n" . join("", map { "\t$_\n" } @_bad_arguments);
+	    my $msg = "Invalid arguments passed to KButil_hide_ws_objects:\n" . join("", map { "\t$_\n" } @_bad_arguments);
 	    Bio::KBase::Exceptions::ArgumentValidationError->throw(error => $msg,
-								   method_name => 'KButil_delete_ws_objects');
+								   method_name => 'KButil_hide_ws_objects');
 	}
     }
 
     my $url = $self->{url};
     my $result = $self->{client}->call($url, $self->{headers}, {
-	    method => "kb_ObjectUtilities.KButil_delete_ws_objects",
+	    method => "kb_ObjectUtilities.KButil_hide_ws_objects",
 	    params => \@args,
     });
     if ($result) {
 	if ($result->is_error) {
 	    Bio::KBase::Exceptions::JSONRPC->throw(error => $result->error_message,
 					       code => $result->content->{error}->{code},
-					       method_name => 'KButil_delete_ws_objects',
+					       method_name => 'KButil_hide_ws_objects',
 					       data => $result->content->{error}->{error} # JSON::RPC::ReturnObject only supports JSONRPC 1.1 or 1.O
 					      );
 	} else {
 	    return wantarray ? @{$result->result} : $result->result->[0];
 	}
     } else {
-        Bio::KBase::Exceptions::HTTP->throw(error => "Error invoking method KButil_delete_ws_objects",
+        Bio::KBase::Exceptions::HTTP->throw(error => "Error invoking method KButil_hide_ws_objects",
 					    status_line => $self->{client}->status_line,
-					    method_name => 'KButil_delete_ws_objects',
+					    method_name => 'KButil_hide_ws_objects',
 				       );
     }
 }
  
 
 
-=head2 KButil_undelete_ws_objects
+=head2 KButil_unhide_ws_objects
 
-  $return = $obj->KButil_undelete_ws_objects($params)
+  $return = $obj->KButil_unhide_ws_objects($params)
 
 =over 4
 
@@ -441,16 +441,16 @@ data_obj_ref is a string
 =begin html
 
 <pre>
-$params is a kb_ObjectUtilities.KButil_undelete_ws_objects_Params
-$return is a kb_ObjectUtilities.KButil_undelete_ws_objects_Output
-KButil_undelete_ws_objects_Params is a reference to a hash where the following keys are defined:
+$params is a kb_ObjectUtilities.KButil_unhide_ws_objects_Params
+$return is a kb_ObjectUtilities.KButil_unhide_ws_objects_Output
+KButil_unhide_ws_objects_Params is a reference to a hash where the following keys are defined:
 	workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 	object_types has a value which is a reference to a list where each element is a string
 	verbose has a value which is a kb_ObjectUtilities.bool
-	undelete_all has a value which is a kb_ObjectUtilities.bool
+	unhide_all has a value which is a kb_ObjectUtilities.bool
 workspace_name is a string
 bool is an int
-KButil_undelete_ws_objects_Output is a reference to a hash where the following keys are defined:
+KButil_unhide_ws_objects_Output is a reference to a hash where the following keys are defined:
 	report_name has a value which is a kb_ObjectUtilities.data_obj_name
 	report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 data_obj_name is a string
@@ -462,16 +462,16 @@ data_obj_ref is a string
 
 =begin text
 
-$params is a kb_ObjectUtilities.KButil_undelete_ws_objects_Params
-$return is a kb_ObjectUtilities.KButil_undelete_ws_objects_Output
-KButil_undelete_ws_objects_Params is a reference to a hash where the following keys are defined:
+$params is a kb_ObjectUtilities.KButil_unhide_ws_objects_Params
+$return is a kb_ObjectUtilities.KButil_unhide_ws_objects_Output
+KButil_unhide_ws_objects_Params is a reference to a hash where the following keys are defined:
 	workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 	object_types has a value which is a reference to a list where each element is a string
 	verbose has a value which is a kb_ObjectUtilities.bool
-	undelete_all has a value which is a kb_ObjectUtilities.bool
+	unhide_all has a value which is a kb_ObjectUtilities.bool
 workspace_name is a string
 bool is an int
-KButil_undelete_ws_objects_Output is a reference to a hash where the following keys are defined:
+KButil_unhide_ws_objects_Output is a reference to a hash where the following keys are defined:
 	report_name has a value which is a kb_ObjectUtilities.data_obj_name
 	report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 data_obj_name is a string
@@ -488,7 +488,7 @@ data_obj_ref is a string
 
 =cut
 
- sub KButil_undelete_ws_objects
+ sub KButil_unhide_ws_objects
 {
     my($self, @args) = @_;
 
@@ -497,7 +497,7 @@ data_obj_ref is a string
     if ((my $n = @args) != 1)
     {
 	Bio::KBase::Exceptions::ArgumentValidationError->throw(error =>
-							       "Invalid argument count for function KButil_undelete_ws_objects (received $n, expecting 1)");
+							       "Invalid argument count for function KButil_unhide_ws_objects (received $n, expecting 1)");
     }
     {
 	my($params) = @args;
@@ -505,31 +505,31 @@ data_obj_ref is a string
 	my @_bad_arguments;
         (ref($params) eq 'HASH') or push(@_bad_arguments, "Invalid type for argument 1 \"params\" (value was \"$params\")");
         if (@_bad_arguments) {
-	    my $msg = "Invalid arguments passed to KButil_undelete_ws_objects:\n" . join("", map { "\t$_\n" } @_bad_arguments);
+	    my $msg = "Invalid arguments passed to KButil_unhide_ws_objects:\n" . join("", map { "\t$_\n" } @_bad_arguments);
 	    Bio::KBase::Exceptions::ArgumentValidationError->throw(error => $msg,
-								   method_name => 'KButil_undelete_ws_objects');
+								   method_name => 'KButil_unhide_ws_objects');
 	}
     }
 
     my $url = $self->{url};
     my $result = $self->{client}->call($url, $self->{headers}, {
-	    method => "kb_ObjectUtilities.KButil_undelete_ws_objects",
+	    method => "kb_ObjectUtilities.KButil_unhide_ws_objects",
 	    params => \@args,
     });
     if ($result) {
 	if ($result->is_error) {
 	    Bio::KBase::Exceptions::JSONRPC->throw(error => $result->error_message,
 					       code => $result->content->{error}->{code},
-					       method_name => 'KButil_undelete_ws_objects',
+					       method_name => 'KButil_unhide_ws_objects',
 					       data => $result->content->{error}->{error} # JSON::RPC::ReturnObject only supports JSONRPC 1.1 or 1.O
 					      );
 	} else {
 	    return wantarray ? @{$result->result} : $result->result->[0];
 	}
     } else {
-        Bio::KBase::Exceptions::HTTP->throw(error => "Error invoking method KButil_undelete_ws_objects",
+        Bio::KBase::Exceptions::HTTP->throw(error => "Error invoking method KButil_unhide_ws_objects",
 					    status_line => $self->{client}->status_line,
-					    method_name => 'KButil_undelete_ws_objects',
+					    method_name => 'KButil_unhide_ws_objects',
 				       );
     }
 }
@@ -1264,7 +1264,7 @@ ws_obj_refs has a value which is a reference to a hash where the key is a string
 
 
 
-=head2 KButil_delete_ws_objects_Params
+=head2 KButil_hide_ws_objects_Params
 
 =over 4
 
@@ -1272,9 +1272,9 @@ ws_obj_refs has a value which is a reference to a hash where the key is a string
 
 =item Description
 
-KButil_delete_ws_objects()
+KButil_hide_ws_objects()
 **
-**  Method for deleting workspace objects
+**  Method for hiding workspace objects in bulk
 
 
 =item Definition
@@ -1286,7 +1286,7 @@ a reference to a hash where the following keys are defined:
 workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 object_types has a value which is a reference to a list where each element is a string
 verbose has a value which is a kb_ObjectUtilities.bool
-delete_all has a value which is a kb_ObjectUtilities.bool
+hide_all has a value which is a kb_ObjectUtilities.bool
 
 </pre>
 
@@ -1298,7 +1298,7 @@ a reference to a hash where the following keys are defined:
 workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 object_types has a value which is a reference to a list where each element is a string
 verbose has a value which is a kb_ObjectUtilities.bool
-delete_all has a value which is a kb_ObjectUtilities.bool
+hide_all has a value which is a kb_ObjectUtilities.bool
 
 
 =end text
@@ -1307,7 +1307,7 @@ delete_all has a value which is a kb_ObjectUtilities.bool
 
 
 
-=head2 KButil_delete_ws_objects_Output
+=head2 KButil_hide_ws_objects_Output
 
 =over 4
 
@@ -1339,7 +1339,7 @@ report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 
 
 
-=head2 KButil_undelete_ws_objects_Params
+=head2 KButil_unhide_ws_objects_Params
 
 =over 4
 
@@ -1347,9 +1347,9 @@ report_ref has a value which is a kb_ObjectUtilities.data_obj_ref
 
 =item Description
 
-KButil_undelete_ws_objects()
+KButil_unhide_ws_objects()
 **
-**  Method for undeleting workspace objects
+**  Method for unhiding workspace objects in bulk
 
 
 =item Definition
@@ -1361,7 +1361,7 @@ a reference to a hash where the following keys are defined:
 workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 object_types has a value which is a reference to a list where each element is a string
 verbose has a value which is a kb_ObjectUtilities.bool
-undelete_all has a value which is a kb_ObjectUtilities.bool
+unhide_all has a value which is a kb_ObjectUtilities.bool
 
 </pre>
 
@@ -1373,7 +1373,7 @@ a reference to a hash where the following keys are defined:
 workspace_name has a value which is a kb_ObjectUtilities.workspace_name
 object_types has a value which is a reference to a list where each element is a string
 verbose has a value which is a kb_ObjectUtilities.bool
-undelete_all has a value which is a kb_ObjectUtilities.bool
+unhide_all has a value which is a kb_ObjectUtilities.bool
 
 
 =end text
@@ -1382,7 +1382,7 @@ undelete_all has a value which is a kb_ObjectUtilities.bool
 
 
 
-=head2 KButil_undelete_ws_objects_Output
+=head2 KButil_unhide_ws_objects_Output
 
 =over 4